--- src/components/catalog-footer-first-row/index.ts
@@ -1,1 +1,0 @@
-export { default } from './catalog-footer-first-row';
