--- src/components/footer-first-row/footer-first-row.module.css
@@ -1,27 +1,0 @@
-/* Shared footer first-row layout styles (catalog footer, board pagination footer mode) */
-.footerRow {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  gap: 8px;
-  width: 100%;
-}
-
-.footerLeft {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  flex-wrap: wrap;
-}
-
-.footerRight {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-}
-
-.styleLabel {
-  font-size: 12px;
-  text-transform: capitalize;
-}
