+++ src/components/footer/index.ts
@@ -1,0 +1,2 @@
+export { PageFooterDesktop, CatalogFooterFirstRow, ThreadFooterFirstRow } from './footer';
+export type { PageFooterDesktopProps, ThreadFooterFirstRowProps } from './footer';
