--- src/components/page-footer-desktop/index.ts
@@ -1,2 +1,0 @@
-export { default } from './page-footer-desktop';
-export type { PageFooterDesktopProps } from './page-footer-desktop';
