--- src/components/thread-footer-first-row/index.ts
@@ -1,2 +1,0 @@
-export { default } from './thread-footer-first-row';
-export type { ThreadFooterFirstRowProps } from './thread-footer-first-row';
