--- src/components/thread-footer-first-row/thread-footer-first-row.module.css
@@ -1,31 +1,0 @@
-.row {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  gap: 12px;
-  width: 100%;
-  text-transform: capitalize;
-}
-
-.left {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  flex-wrap: wrap;
-}
-
-.center {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.right {
-  display: flex;
-  align-items: center;
-}
-
-.row .button {
-  text-transform: capitalize;
-}
