+++ src/views/catalog/catalog.tsx
@@ -17,10 +17,9 @@ import useSortingStore from '../../stores/use-sorting-store';
 import useCatalogFiltersStore from '../../stores/use-catalog-filters-store';
 import { getSubplebbitAddress, isDirectoryBoard, normalizeMultiboardFeedPath } from '../../lib/utils/route-utils';
 import CatalogRow from '../../components/catalog-row';
-import CatalogFooterFirstRow from '../../components/catalog-footer-first-row';
+import { CatalogFooterFirstRow, PageFooterDesktop } from '../../components/footer';
 import LoadingEllipsis from '../../components/loading-ellipsis';
 import ErrorDisplay from '../../components/error-display/error-display';
-import PageFooterDesktop from '../../components/page-footer-desktop';
 import styles from './catalog.module.css';
 import { commentMatchesPattern } from '../../lib/utils/pattern-utils';
 import { sortCatalogFeedForDisplay } from '../../lib/utils/catalog-sort';
